14-binary_tree_balance

+++ 14-binary_tree_balance.c
@@ -9,7 +9,7 @@
 
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	int left_height, right_height = 0;
+	size_t left_height, right_height = 0;
 
 	left_height = binary_tree_height(tree->left);
 	right_height = binary_tree_height(tree->right);
